First optimization
Made a hash map of the hash values of characters. If a character is encountered for the first time, the hash is calculated and stored in the hash map. If it has already been encountered, can simply read its hash value from the hash map. Runtime is now over 10x shorter.
Updated intial-graph.html, as it hadn't been updating on some runs because of the typo in the last commit.